+++ packages/access-control/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@iguhealth/access-control",
+  "version": "0.0.1",
+  "type": "module",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "scripts": {
+    "build": "tsc",
+    "publish": "yarn build && yarn npm publish --access public --tolerate-republish"
+  },
+  "dependencies": {
+    "@iguhealth/client": "workspace:^",
+    "@iguhealth/fhir-types": "workspace:^",
+    "@iguhealth/jwt": "workspace:^",
+    "@iguhealth/operation-outcomes": "workspace:^"
+  },
+  "devDependencies": {
+    "@jest/globals": "^29.7.0",
+    "@types/jest": "^29.5.12",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.4",
+    "typescript": "5.5.4"
+  }
+}
