+++ packages/access-control/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../config/base.tsconfig.json",
+  "exclude": ["./**/*.test.ts", "lib/*", "jest.config.js"],
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "lib"
+  }
+}
